chore: fix lint errors across src/ and tests/
Brought oxlint down from 79 errors to 0 (32 cosmetic warnings remain).

- Exclude design/ prototypes from lint — they're reference HTML/JSX, not
  app code, and were the source of most of the errors.
- Extract ApprovalEyebrow from ProgramApproval and HymnPickerPopover
  from HymnPicker so both files fit under the 150-LOC ceiling.
- Mechanical fixes: .sort() → .toSorted(), drop unused imports/vars,
  rename a shadowed prev binding, remove a useless `?? {}` fallback
  in an object spread.

No behavior changes.
