+++ .oxlintrc.json
@@ -60,5 +60,5 @@
       }
     }
   ],
-  "ignorePatterns": ["dist", "node_modules", "playwright-report", "test-results", "coverage"]
+  "ignorePatterns": ["dist", "node_modules", "playwright-report", "test-results", "coverage", "design"]
 }
